+++ packages/terra-badge/CHANGELOG.md
@@ -3,7 +3,9 @@
 ## Unreleased
 
 * Added
-  * Added a11y changes to terra-badge
+  * Added hidden text to announce the intent of the badge
+* Changed
+  * Removed `aria-hidden` from the textContent to announce text of badge
 
 ## 3.57.0 - (February 15, 2023)